+++ shared/src/commonMain/kotlin/ly/david/musicsearch/shared/image/imageLoader.common.kt
@@ -33,7 +33,7 @@ fun newImageLoader(
         .build()
 }
 
-internal fun newDiskCache(): DiskCache? {
+internal fun newDiskCache(): DiskCache {
     return DiskCache.Builder()
         .directory(FileSystem.SYSTEM_TEMPORARY_DIRECTORY / "image_cache")
         .maxSizeBytes(512L * 1024 * 1024) // 512MB
