+++ ui/image/build.gradle.kts
@@ -27,17 +27,7 @@ kotlin {
         val androidMain by getting {
             dependencies {
                 implementation(compose.preview)
-//                implementation(libs.coil.network.okhttp)
             }
-        }
-
-        jvmMain.dependencies {
-//            implementation(libs.coil.network.okhttp)
-
-        }
-
-        iosMain.dependencies {
-//            implementation("io.ktor:ktor-client-darwin:3.0.0-wasm2")
         }
     }
 }
